Stats: do not multiply int and double
Convert int to double before the multiplication.

Example of wrong pangenome.info generated:

Exact stem blocks (represented in all genomes) but not minor:
fragments / blocks = 1624 / 203 = 8
Block identity: number=203 min=0.9527 median=0.9957 avg=0.9943 max=1
Identity of joined blocks: 0.0320695

(See the last line.)

+++ src/algo/Stats.cpp
@@ -67,7 +67,8 @@ static void report_weighted_average_identity(
         AlignmentStat al_stat;
         make_stat(al_stat, block);
         double identity = block_identity(al_stat).to_d();
-        identity_wsum += block->alignment_length() * identity;
+        identity_wsum += double(block->alignment_length()) *
+            identity;
         length_sum += block->alignment_length();
     }
     if (!bs->empty()) {
